Fixed poorly written variable setup
Fixed an issue with the variable being setup. The issue being that there was a declaration to empty the variable, then (in a below 'if' statement) to check if that variable was empty (which it always would be). The variables are now specifically set as empty, only if they are not set. 
This may need a later update with the '-z $VARIABLE' to check if it exists, but a problem for another day.