Adding a default configuration file
There are a few settings (which appear not to be configured for the auto launch script), in the use-case I have, thus they have been included in this file. The most relevant one is the disabling discovery, which for the Home Assistant purpose doesn't actually matter anyway (as the API which raspotify doesn't allow for this anyway). It does mean that USER and PASS are required when running the container, as that is the only way for the device to appear (again, if HA was able to use it without this, this whole endevour would not be required).